Tested API for databases, all working
Changes to be committed:
	modified:   TODO
	modified:   data/database.db
	modified:   src/crud.py
	modified:   src/main.py
	modified:   src/models.py
	modified:   src/schemas.py

+++ src/main.py
@@ -7,8 +7,10 @@ import crud
 import models
 import schemas
 from database import SessionLocal, engine
+import constants
 
-models.Base.metadata.drop_all(engine)
+if constants.debug:
+    models.Base.metadata.drop_all(engine)
 models.Base.metadata.create_all(bind=engine)
 
 
@@ -28,7 +30,7 @@ def root():
     return {'Hello': 'World'}
 
 
-@app.post('/add_device/{house_id}', response_model=schemas.Device)
+@app.post('/add_device/{house_id}', response_model=schemas.Device)  # OK
 def add_device(house_id: int, device: schemas.DeviceCreate, db: Session = Depends(get_db)):
     # Add device, return created id
     db_device = crud.get_device_by_sn(db, device.serial_number)
@@ -37,7 +39,7 @@ def add_device(house_id: int, device: schemas.DeviceCreate, db: Session = Depend
     return crud.create_house_device(db, device, house_id)
 
 
-@app.delete('/del_device/{device_id}')
+@app.delete('/del_device/{device_id}')  # OK
 def del_device(device_id: int, db: Session = Depends(get_db)):
     # Delete device and return status
     result = crud.delete_device(db, device_id)
@@ -46,7 +48,7 @@ def del_device(device_id: int, db: Session = Depends(get_db)):
     return {"status": "OK"}
 
 
-@app.post('/add_user/', response_model=schemas.User)
+@app.post('/add_user/', response_model=schemas.User)  # OK
 def add_user(user: schemas.UserCreate, db: Session = Depends(get_db)):
     # Add user and return unique code
     db_user = crud.get_user_by_email(db, email=user.email)
@@ -55,13 +57,13 @@ def add_user(user: schemas.UserCreate, db: Session = Depends(get_db)):
     return crud.create_user(db=db, user=user)
 
 
-@app.get("/get_users/", response_model=List[schemas.User])
+@app.get("/get_users/", response_model=List[schemas.User])  # OK
 def read_users(skip:int = 0, limit: int = 100, db: Session = Depends(get_db)):
     users = crud.get_users(db, skip=skip, limit=limit)
     return users
 
 
-@app.get('/get_user/{email}', response_model=schemas.User)
+@app.get('/get_user/{email}', response_model=schemas.User)  # OK
 def get_user(email: str, db: Session = Depends(get_db)):
     # Check if there is user
     db_user = crud.get_user_by_email(db, email=email)
@@ -76,13 +78,15 @@ def operate_device(action: schemas.DeviceAction):
     pass
 
 
-@app.post('/add_rule/{house_id}', response_model=schemas.Rule)
+@app.post('/add_rule/{house_id}', response_model=schemas.Rule)  # OK
 def add_rule(house_id: int, rule: schemas.RuleCreate, db: Session = Depends(get_db)):
     # Add the rule
-    return crud.create_house_rule(db, rule=rule, house_id=house_id)
+    x = crud.create_house_rule(db, rule=rule, house_id=house_id)
+    print("OK")
+    return x
 
 
-@app.delete('/del_rule/{rule_id}')
+@app.delete('/del_rule/{rule_id}')  # OK
 def del_rule(rule_id: int, db: Session = Depends(get_db)):
     res = crud.delete_rule(db, rule_id)
     if res < 0:
@@ -90,7 +94,7 @@ def del_rule(rule_id: int, db: Session = Depends(get_db)):
     return {"status": "OK"}
 
 
-@app.get('/get_rules/{house_id}', response_model=List[schemas.Rule])
+@app.get('/get_rules/{house_id}', response_model=List[schemas.Rule])  # OK
 def get_rules(house_id: int, db: Session = Depends(get_db)):
     # Get rules
     db_rules = crud.get_rules_by_house(db, house_id)
@@ -105,7 +109,7 @@ def set_schedule(device_id: int, schedule: schemas.ScheduleItem):
     pass
 
 
-@app.get('/get_devices/{house_id}', response_model=List[schemas.Device])
+@app.get('/get_devices/{house_id}', response_model=List[schemas.Device])  # OK
 def get_devices(house_id: int, db: Session = Depends(get_db)):
     # Get a list of devices
     db_devices = crud.get_devices_by_house(db, house_id)
@@ -114,7 +118,7 @@ def get_devices(house_id: int, db: Session = Depends(get_db)):
     return db_devices
 
 
-@app.get('/get_houses/{user_id}', response_model=List[schemas.User])
+@app.get('/get_houses/{user_id}', response_model=List[schemas.House])  # OK
 def get_houses(user_id: int, db: Session = Depends(get_db)):
     # Get houses
     db_houses = crud.get_houses_by_owner(db, user_id)
@@ -123,13 +127,15 @@ def get_houses(user_id: int, db: Session = Depends(get_db)):
     return db_houses
 
 
-@app.post('/add_house/{user_id}', response_model=schemas.House)
-def add_house(user_id: int, house: schemas.HouseCreate, db: Session = Depends(get_db)):
-    # Mobile app responsibility to check if house exists.
-    return crud.create_user_house(db=db, house=house, user_id=user_id)
+@app.post('/add_house/{user_id}', response_model=schemas.House)  # OK
+def add_house(user_id: int, house: schemas.HouseCreate, db: Session = Depends(get_db)):  # OK
+    new_house = crud.create_user_house(db=db, house=house, user_id=user_id)
+    if new_house is None:
+        raise HTTPException(status_code=400, detail="House with same name already added")
+    return new_house
 
 
-@app.delete('/del_house/{house_id}')
+@app.delete('/del_house/{house_id}')  # OK
 def del_house(house_id: int, db: Session = Depends(get_db)):
     res = crud.delete_house(db, house_id)
     if res < 0:
@@ -137,7 +143,7 @@ def del_house(house_id: int, db: Session = Depends(get_db)):
     return {"status": "OK"}
 
 
-@app.post('/update_rule/{house_id}/{rule_id}', response_model=schemas.Rule)
+@app.post('/update_rule/{house_id}/{rule_id}', response_model=schemas.Rule)  # OK
 def update_rule(house_id: int, rule_id: int, new_rule: schemas.RuleCreate, db: Session = Depends(get_db)):
     # Delete rule and add
     crud.delete_rule(db, rule_id)
